backend/requirements.txt
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.6.0
django-rest-auth==0.9.5
django-filter
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
httpie==2.4.0
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
requests==2.25.1
djangorestframework-jwt

dj3-cloudinary-storage

gunicorn
django-storages[azure]
